requirements.txt
anyio==3.2.1
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.4.0
astroid==2.4.2
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bcrypt==3.2.0
black==21.6b0
bleach==3.3.0
CacheControl==0.12.6
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
click==8.0.1
colorama==0.4.3
cycler==0.11.0
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.10.1
django-heroku==0.3.1
djangorestframework==3.12.4
entrypoints==0.3
et-xmlfile==1.1.0
firebase-admin==5.0.3
google-api-core==2.1.0
google-api-python-client==2.24.0
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-cloud-core==2.1.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
googletrans==3.0.0
grpcio==1.41.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.10.1
httpcore==0.9.1
httplib2==0.20.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
kiwisolver==1.3.2
lazy-object-proxy==1.4.3
libretranslatepy==2.1.1
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.25
mysqlclient==2.0.3
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.0
openpyxl==3.0.9
packaging==21.0
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
paytmchecksum==1.7.0
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.19
proto-plus==1.19.4
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.11.0
Pygments==2.9.0
pylint==2.5.3
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
pyrsistent==0.18.0
python-dateutil==2.8.1
python-decouple==3.5
pytz==2021.1
pyzmq==22.1.0
regex==2021.7.6
requests==2.25.1
requests-unixsocket==0.2.0
rfc3986==1.5.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.7.1
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
terminado==0.9.5
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.1
tornado==6.1
traitlets==5.0.5
translate==3.6.1
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
whitenoise==5.2.0
wrapt==1.12.1
xlrd==2.0.1
